refactor(subscriptions): move product helpers to integration class
